+++ src/rules/function-comma-space-before/__tests__/index.js
@@ -99,3 +99,59 @@ testRule("never", tr => {
     column: 46,
   })
 })
+
+testRule("always-single-line", tr => {
+  warningFreeBasics(tr)
+
+  tr.ok("a::before { content: \"func(foo,bar,baz)\"; }")
+  tr.ok("a::before { background: url('func(foo,bar,baz)'); }")
+  tr.ok("a { background-size: 0, 0, 0; }")
+  tr.ok("a { transform: translate(1 , 1); }")
+  tr.ok("a { transform: translate(1 ,1); }")
+  tr.ok("a { transform: color(rgb(0 , 0 ,0) lightness(50%)); }")
+
+  tr.ok("a { transform: translate(1,\n1); }")
+  tr.ok("a { transform: translate(1  ,\n1); }")
+  tr.ok("a { transform: translate(1\t,\r\n1); }", "CRLF")
+  tr.ok("a { transform: translate(1\n, 1); }")
+  tr.ok("a { transform: translate(1\n,\n1); }")
+
+  tr.notOk("a { transform: color(rgb(0 , 0, 0) lightness(50%)); }", {
+    message: messages.expectedBeforeSingleLine(),
+    line: 1,
+    column: 31,
+  })
+  tr.notOk("a { transform: color(lightness(50%) rgb(0 , 0, 0)); }", {
+    message: messages.expectedBeforeSingleLine(),
+    line: 1,
+    column: 46,
+  })
+})
+
+testRule("never-single-line", tr => {
+  warningFreeBasics(tr)
+
+  tr.ok("a::before { content: \"func(foo ,bar ,baz)\"; }")
+  tr.ok("a::before { background: url('func(foo ,bar ,baz)'); }")
+  tr.ok("a { background-size: 0 , 0 , 0; }")
+  tr.ok("a { transform: translate(1, 1); }")
+  tr.ok("a { transform: translate(1,1); }")
+  tr.ok("a { transform: color(rgb(0, 0,0) lightness(50%)); }")
+
+  tr.ok("a { transform: translate(1 ,\n1); }")
+  tr.ok("a { transform: translate(1  ,\n1); }")
+  tr.ok("a { transform: translate(1\t,\r\n1); }", "CRLF")
+  tr.ok("a { transform: translate(1\n, 1); }")
+  tr.ok("a { transform: translate(1\n,\n1); }")
+
+  tr.notOk("a { transform: color(rgb(0, 0 , 0) lightness(50%)); }", {
+    message: messages.rejectedBeforeSingleLine(),
+    line: 1,
+    column: 31,
+  })
+  tr.notOk("a { transform: color(lightness(50%) rgb(0, 0 , 0)); }", {
+    message: messages.rejectedBeforeSingleLine(),
+    line: 1,
+    column: 46,
+  })
+})