+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 * Added: support for using `.stylelintrc` JSON file.
 * Added: support for extending existing configs using the `extends` property.
 * Added: support for SCSS syntax parsing to CLI and Node API.
+* Added: `function-comma-newline-after` rule.
+* Added: `function-comma-newline-before` rule.
+* Added: `"always-single-line"` and `"never-single-line"` options to `function-comma-space-after` rule.
+* Added: `"always-single-line"` and `"never-single-line"` options to `function-comma-space-before` rule.
 
 # 1.2.1